@@ -4,10 +4,13 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.2.2",
-    "@craco/craco": "^5.6.4",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.58",
+    "@mui/material": "^5.11.2",
+    "@rollup/plugin-babel": "^6.0.3",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
@@ -21,9 +24,9 @@
     "prop-types": "^15.7.2",
     "query-string": "^6.13.1",
     "rc-color-picker": "^1.2.6",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "react-headroom": "^3.0.0",
     "react-image": "^4.0.3",
     "react-image-crop": "^8.6.12",
@@ -33,7 +36,6 @@
     "react-responsive": "^8.1.0",
     "react-reveal": "^1.2.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
     "react-social-login-buttons": "^3.4.0",
     "react-syntax-highlighter": "^15.4.3",
     "redux": "^4.0.5",
@@ -43,12 +45,12 @@
     "validator": "^13.6.0"
   },
   "scripts": {
-    "start": "craco start",
+    "start": "vite",
     "emulator": "firebase emulators:start",
     "emulator-import": "firebase emulators:start --import=testdata",
     "emulator-export": "rm -rf testdata && firebase emulators:export testdata",
-    "build": "craco build",
-    "test": "craco test",
+    "build": "vite build",
+    "test": "vite test",
     "eject": "react-scripts eject",
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public",
@@ -76,10 +78,12 @@
     "@storybook/node-logger": "^6.2.9",
     "@storybook/preset-create-react-app": "^3.1.7",
     "@storybook/react": "^6.2.9",
+    "@vitejs/plugin-react": "1.1.1",
     "css-loader": "^5.2.6",
     "cypress": "^4.12.1",
     "cypress-xpath": "^1.6.0",
     "less-loader": "^7.3.0",
-    "style-loader": "^2.0.0"
+    "style-loader": "^2.0.0",
+    "vite": "2.9.0"
   }
 }
