@@ -9,14 +9,14 @@ import "firebase/performance";
 import "firebase/messaging";
 
 const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
-  databaseURL: `https://${process.env.REACT_APP_FIREBASE_PROJECT_ID}.firebaseio.com`,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: `${process.env.REACT_APP_FIREBASE_PROJECT_ID}.appspot.com`,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID,
-  measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENTID,
+  apiKey: import.meta.env.VITE_APP_FIREBASE_API_KEY,
+  authDomain: import.meta.env.VITE_APP_AUTH_DOMAIN,
+  databaseURL: `https://${import.meta.env.VITE_APP_FIREBASE_PROJECT_ID}.firebaseio.com`,
+  projectId: import.meta.env.VITE_APP_FIREBASE_PROJECT_ID,
+  storageBucket: `${import.meta.env.VITE_APP_FIREBASE_PROJECT_ID}.appspot.com`,
+  messagingSenderId: import.meta.env.VITE_APP_FIREBASE_MESSAGING_SENDER_ID,
+  appId: import.meta.env.VITE_APP_FIREBASE_APP_ID,
+  measurementId: import.meta.env.VITE_APP_FIREBASE_MEASUREMENTID,
 };
 
 //console.log("firebaseConfig", firebaseConfig);
@@ -48,7 +48,7 @@ let firebase_messaging;
 if (firebase.messaging.isSupported()) {
   firebase_messaging = firebase.messaging();
   firebase_messaging.usePublicVapidKey(
-    process.env.REACT_APP_FIREBASE_FCM_VAPID_KEY
+    import.meta.env.VITE_APP_FIREBASE_FCM_VAPID_KEY
   );
 }
 
